requirements.txt
Django>=5.0
djangorestframework
gunicorn
psycopg2-binary
django-cors-headers
python-dotenv
